requirements.txt
amqp==5.3.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
billiard==4.2.4
black==25.11.0
celery==5.6.0
class-registry==2.1.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
decorator==5.2.1
Django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
exceptiongroup==1.3.1
executing==2.2.1
filters==1.3.2
flower==2.0.1
gunicorn==23.0.0
humanize==4.14.0
inflection==0.5.1
ipdb==0.13.13
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isort==7.0.0
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
matplotlib-inline==0.2.1
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.4
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
